feat(payments): move totals to the dto
